# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/initiatives": {
      "get": {
        "operationId": "InitiativesController_list",
        "parameters": [
          {
            "name": "kind",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "year",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "quarter",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "is_backlog",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "initiatives"
        ]
      },
      "post": {
        "operationId": "InitiativesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInitiativeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessDto"
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "initiatives"
        ]
      }
    },
    "/api/v1/initiatives/years/{id}": {
      "get": {
        "operationId": "InitiativesController_getYear",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "initiatives"
        ]
      },
      "delete": {
        "operationId": "InitiativesController_deleteYear",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "revision",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "initiatives"
        ]
      }
    },
    "/api/v1/initiatives/cards/{id}": {
      "get": {
        "operationId": "InitiativesController_getCard",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "initiatives"
        ]
      },
      "patch": {
        "operationId": "InitiativesController_updateCard",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCardDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "initiatives"
        ]
      },
      "delete": {
        "operationId": "InitiativesController_deleteCard",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "revision",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "initiatives"
        ]
      }
    },
    "/api/v1/initiatives/years/{id}/cards": {
      "post": {
        "operationId": "InitiativesController_createCard",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuarterCardDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessDto"
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "initiatives"
        ]
      }
    },
    "/api/v1/initiatives/cards/{id}/move": {
      "post": {
        "operationId": "InitiativesController_moveCard",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PeriodCommandDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessDto"
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "initiatives"
        ]
      }
    },
    "/api/v1/initiatives/cards/{id}/continue": {
      "post": {
        "operationId": "InitiativesController_continueCard",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PeriodCommandDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessDto"
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "initiatives"
        ]
      }
    },
    "/api/v1/initiatives/cards/{cardId}/scope/{itemId}/move": {
      "post": {
        "operationId": "InitiativesController_moveScope",
        "parameters": [
          {
            "name": "cardId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PeriodCommandDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessDto"
                }
              }
            }
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "initiatives"
        ]
      }
    },
    "/api/v1/initiatives/years/{id}/passport": {
      "post": {
        "operationId": "InitiativesController_saveYearPassport",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavePassportDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessDto"
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "initiatives"
        ]
      }
    },
    "/api/v1/initiatives/cards/{id}/passport": {
      "post": {
        "operationId": "InitiativesController_saveCardPassport",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavePassportDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessDto"
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "initiatives"
        ]
      }
    },
    "/api/v1/initiatives/years/extend": {
      "post": {
        "operationId": "InitiativesController_extend",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtendYearsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessDto"
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "initiatives"
        ]
      }
    },
    "/api/v1/initiatives/years/{id}/preparation": {
      "patch": {
        "operationId": "InitiativesController_updatePreparation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePreparationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "initiatives"
        ]
      }
    },
    "/api/v1/users": {
      "get": {
        "operationId": "UsersController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "users"
        ]
      },
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessDto"
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/{id}": {
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/{id}/reset-password": {
      "post": {
        "operationId": "UsersController_resetPassword",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessDto"
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "users"
        ]
      }
    },
...
  },
  "info": {
    "title": "PMO Hub API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ApiSuccessDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true,
            "enum": [
              true
            ]
          },
          "message": {
            "type": "string",
            "example": "Операцію виконано"
          },
          "data": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "success"
        ]
      },
      "PassportDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "strategic_goal": {
            "type": "string"
          },
          "manager_id": {
            "type": "string",
            "format": "uuid"
          },
          "priority": {
            "type": "string",
            "format": "uuid"
          },
          "notes": {
            "type": "string"
          },
          "implementer_dept_ids": {
            "default": [],
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "cross_functional_dept_ids": {
            "default": [],
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "custom_fields": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "name",
          "implementer_dept_ids",
          "cross_functional_dept_ids"
        ]
      },
      "ChecklistItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "text": {
            "type": "string"
          },
          "is_completed": {
            "type": "boolean"
          },
          "color": {
            "type": "string"
          },
          "weightId": {
            "type": "string",
            "format": "uuid"
          },
          "weightSnapshot": {
            "type": "object",
            "properties": {
              "definitionId": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "value": {
                "type": "number"
              }
            },
            "required": [
              "name",
              "value"
            ]
          },
          "assigneeIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "implementer_dept_ids": {
            "default": [],
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "text",
          "implementer_dept_ids"
        ]
      },
      "CreateInitiativeDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "project",
              "task",
              "PROJECT",
              "TASK"
            ]
          },
          "year": {
            "type": "number",
            "minimum": 2000,
            "maximum": 2200
          },
          "passport": {
            "$ref": "#/components/schemas/PassportDto"
          },
          "quarters": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "Q1",
                "Q2",
                "Q3",
                "Q4"
              ]
            }
          },
          "initial_scope": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChecklistItemDto"
            }
          }
        },
        "required": [
          "kind",
          "year",
          "passport",
          "quarters"
        ]
      },
      "CreateQuarterCardDto": {
        "type": "object",
        "properties": {
          "quarter": {
            "enum": [
              "Q1",
              "Q2",
              "Q3",
              "Q4"
            ],
            "type": "string"
          },
          "passport": {
            "$ref": "#/components/schemas/PassportDto"
          },
          "initial_scope": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChecklistItemDto"
            }
          }
        },
        "required": [
          "quarter",
          "passport"
        ]
      },
      "UpdateCardDto": {
        "type": "object",
        "properties": {
          "revision": {
            "type": "number",
            "minimum": 1
          },
          "passport": {
            "$ref": "#/components/schemas/PassportDto"
          },
          "health_status": {
            "type": "string",
            "format": "uuid"
          },
          "checklist": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChecklistItemDto"
            }
          }
        },
        "required": [
          "revision"
        ]
      },
      "PeriodCommandDto": {
        "type": "object",
        "properties": {
          "revision": {
            "type": "number",
            "minimum": 1
          },
          "to_year": {
            "type": "number",
            "minimum": 2000,
            "maximum": 2200
          },
          "to_quarter": {
            "enum": [
              "Q1",
              "Q2",
              "Q3",
              "Q4"
            ],
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "confirmation_token": {
            "type": "string"
          }
        },
        "required": [
          "revision",
          "to_year",
          "to_quarter"
        ]
      },
      "RevisionTargetDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "revision": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": [
          "id",
          "revision"
        ]
      },
      "SourceCardPatchDto": {
        "type": "object",
        "properties": {
          "health_status": {
            "type": "string",
            "format": "uuid"
          },
          "checklist": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChecklistItemDto"
            }
          }
        }
      },
      "SavePassportDto": {
        "type": "object",
        "properties": {
          "revision": {
            "type": "number",
            "minimum": 1
          },
          "passport": {
            "$ref": "#/components/schemas/PassportDto"
          },
          "target_years": {
            "default": [],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RevisionTargetDto"
            }
          },
          "target_cards": {
            "default": [],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RevisionTargetDto"
            }
          },
          "source_card_patch": {
            "description": "Changes to the source card that must commit atomically with passport propagation.",
            "allOf": [
              {
                "$ref": "#/components/schemas/SourceCardPatchDto"
              }
            ]
          }
        },
        "required": [
          "revision",
          "passport",
          "target_years",
          "target_cards"
        ]
      },
      "ExtendYearsDto": {
        "type": "object",
        "properties": {
          "source_year_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "target_year": {
            "type": "number",
            "minimum": 2000,
            "maximum": 2200
          }
        },
        "required": [
          "source_year_ids",
          "target_year"
        ]
      },
      "UpdatePreparationDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "strategic_goal": {
            "type": "string"
          },
          "manager_id": {
            "type": "string",
            "format": "uuid"
          },
          "priority": {
            "type": "string",
            "format": "uuid"
          },
          "notes": {
            "type": "string"
          },
          "implementer_dept_ids": {
            "default": [],
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "cross_functional_dept_ids": {
            "default": [],
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "custom_fields": {
            "type": "object",
            "additionalProperties": true
          },
          "revision": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": [
          "name",
          "implementer_dept_ids",
          "cross_functional_dept_ids",
          "revision"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "type": "string",
            "enum": [
              "SUPER_ADMIN",
              "ADMIN",
              "USER"
            ]
          },
          "department_id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "name",
          "email",
          "role"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "type": "string",
            "enum": [
              "SUPER_ADMIN",
              "ADMIN",
              "USER"
            ]
          },
          "department_id": {
            "type": "string",
            "format": "uuid"
          },
          "is_active": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}